requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pyjwt
python-jose
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
cryptography